จัด package ของ AddressPicker ใหม่
