+++ sample/src/main/res/layout/activity_region_picker.xml
@@ -28,7 +28,7 @@
         android:layout_height="wrap_content"
         android:text="@string/hello_world" />
 
-    <th.or.nectec.android.widget.thai.AddressPicker
+    <th.or.nectec.android.widget.thai.addresspicker.AddressPicker
         android:id="@+id/addressView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content" />
