+++ widget/src/main/java/th/or/nectec/android/widget/thai/addresspicker/fragment/RegionListFragment.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package th.or.nectec.android.widget.thai;
+package th.or.nectec.android.widget.thai.addresspicker.fragment;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -28,7 +28,8 @@ import android.widget.Toast;
 
 import java.util.List;
 
-import th.or.nectec.android.widget.thai.repository.StubRegionRepository;
+import th.or.nectec.android.widget.thai.R;
+import th.or.nectec.android.widget.thai.addresspicker.repository.StubRegionRepository;
 import th.or.nectec.domain.thai.address.region.RegionChooser;
 import th.or.nectec.domain.thai.address.region.RegionPresenter;
