+++ widget/src/main/res/layout/fragment_address_list_picker.xml
@@ -20,7 +20,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context="th.or.nectec.android.widget.thai.AddressPickerDialogFragment">
+    tools:context="th.or.nectec.android.widget.thai.addresspicker.AddressPickerDialogFragment">
 
 
     <ListView
