+++ widget/src/main/res/layout/fragment_address_picker_dialog.xml
@@ -21,7 +21,7 @@
     android:layout_height="wrap_content"
     android:orientation="vertical"
     android:paddingTop="16dp"
-    tools:context="th.or.nectec.android.widget.thai.AddressPickerDialogFragment">
+    tools:context="th.or.nectec.android.widget.thai.addresspicker.AddressPickerDialogFragment">
 
     <FrameLayout
         android:id="@+id/container"
